mqtt_collector/requirements.txt
paho-mqtt==1.5.1
requests